bcmp.sh: script for comparing Cr50 code branches
The script builds a Cr50 image from scratch, then by examining the
contents of builed/cr50/RW the script figures out the source files
which were used to build the image and then compares all source files
with a different git branch.

The branch to compare with by default is cros/firmware-cr50-9308.B, if
the user wants to compare to a different branch, the name of the
branch can be passed as command line parameter.

Two git trees are examined, cr50 and tpm2. If the other branch does
not exist in either of the trees or building Cr50 image fails, the
script reports error and exits.

The result of running the script is a set of git diffs for files which
are different between branches.

BUG=b:200823466
TEST=ran the script to compare pre-pvt and mp Cr50 branches, observed
     sensible results.

Signed-off-by: Vadim Bendebury <vbendeb@google.com>
Change-Id: Ic044c2d23758eed1a5573385e903e59ed4328635