system: change sleep_mask to atomic_t type
The sleep_mask is used only with atomic_* functions or for reading,
so change to variable type to atomic_t.

Do the same with idle_disabled.

It shouldn't change the generated code. The only noticed change is for
mchp chip. The compiler uses .w asm extenstion for some instructions in
__idle function in clock.c file.

It will be useful for incoming commits related to modifying atomic_t
caused by a change in Zephyr upstream (atomic_t from int to long).

BUG=b:207082842
TEST=make buildall && zmake testall
BRANCH=main

Signed-off-by: Dawid Niedzwiecki <dn@semihalf.com>
Change-Id: I85d05bb911af973f030cf46d3f534c374ef76936