task: change task events bitmask to atomic_t
The events bitmask is used only with atomic_* functions or for reading,
so change to variable type to atomic_t.

It shouldn't change the generated code.

It will be useful for incoming commits related to modifying atomic_t
caused by a change in Zephyr upstream (atomic_t from int to long).

BUG=b:207082842
TEST=make buildall && zmake testall && build_compare.sh -b all
BRANCH=main

Signed-off-by: Dawid Niedzwiecki <dn@semihalf.com>
Change-Id: Ia0ff4a20f9ed6c3728b1bd8b6cab3b32ebfcc36b