atomic: use atomic_t where it is possible
There are several places where atomic_t can be a type variables that
are use with atomic_* operation, so use it.

It sometimes has an impact on the asm code, but it is not significant.

The change will be useful for incoming commits related to modifying
atomic_t caused by a change in Zephyr upstream (atomic_t from int to
long).

BUG=b:207082842
TEST=make buildall && zmake testall
BRANCH=main

Signed-off-by: Dawid Niedzwiecki <dn@semihalf.com>
Change-Id: I5c7fa6b74b84454b22074a2a00b5f10003ee9843