zephyr: leverage pwm and fan settings from cros_ec it8xxx2_evb
Leverage pwm and fan settings from cros_ec/main:it8xxx2_evb.
And this is convenient to check pwm or fan channel waveform in
the same pin.

Pwm output 324Hz isn't enough to turn on the fan, so set it to 30000Hz.
And we use 324Hz to test pwm maximum output in EC power saving mode.

BRANCH=none
BUG=none
TEST=zmake board it8xxx2_evb

Signed-off-by: Ruibin Chang <Ruibin.Chang@ite.com.tw>
Change-Id: I0e5c023a634f9a936c29b6bec399d86e66edffeb