chip/stm32/usb_spi.c: Advertise endpoint name
The string USB_STR_SPI_NAME clearly was meant for defining the ASCII
label on the USB endpoint used for forwarding of SPI transactions by
servo micro and similar devices.  However, it seems that it was
accidentally left unused.

This CL will add an ASCII label to the SPI endpoint (similar to how
UART and I2C forwarding endpoint have textual labels.)

Also, this CL adds a new CONFIG_USB_SPI_BUFFER_SIZE which can be used
by particular boards to override the default amount of memory reserved
for forwarding between USB and SPI.  That is, the maximum size of a
single SPI transaction that can be forwarded.  (The default of 298 is
too small for the future HyperDebug board, which will serve as similar
purpose to the Google's servo for OpenTitan development.)

Signed-off-by: Jes B. Klinke <jbk@chromium.org>

BUG=b:192262089
TEST=Watch forwarding of 2048 byte transactions on HyperDebug
BRANCH=None

Change-Id: Ic060194c19b0c2f33f6e861252d5fb5e993349b5