zephyr: Split the guybrush build for skyrim
Rename the guybrush project directory to skyrim.

Guybrush becomes a variant of skyrim that we plan to remove once
skyrim hardware is readily available.

Initially I have it set up so that project-specific DTS customizations
can be made in {project_name}.dts, and project-specific Kconfig
customizations can be made in prj_{project_name}.conf.  Right now
these files are both empty for both guybrush and skyrim, but the idea
is that we start to move things into these files that should be
project-specific in the coming weeks.

BUG=b:208515128
BRANCH=none
TEST=compile for both skyrim and guybrush

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: I3c08cf89b7788b4a56fff9fdfa40e8f8a1347cf9