zephyr: pm: npcx: add pm constraint check in PM policy
In PR:40802, the UART driver uses the pm_constraint counter to indicate
if the UART Tx is still ongoing or not. The pm policy should check the
counter with the pm_constraint_get function to determine if the system
can enter the suspend state or stay in the active one.

BUG=none
BRANCH=none
TEST=pass "zmake testall"
TEST=build npcx9_evb and volteer with PR:40802, make sure they can
boot up and enter deep sleep. Make sure the console messages don't
break.

Signed-off-by: Jun Lin <CHLin56@nuvoton.com>
Signed-off-by: Wealian Liao <WHLIAO@nuvoton.com>

Change-Id: I0803d9bb8454825a7ae876abaac58c65e7db1550