drawcia: Don't report HDMI_HPD to CPU when power state S5
Since PP5000_HDMI is enabled by EN_PP5000_U, which means that it's
still turned on when power state S5, VOLUP_BTN_ODL_HDMI_HPD will be
asserted when HDMI is plugged in. If we report this to CPU through
EC_AP_USB_C1_HDMI_HPD, there will be leakage current on CPU side
which might lead to unabling to power on the system.

BUG=none
BRANCH=dedede
TEST=make sure that EC_AP_USB_C1_HDMI_HPD is deasserted when power
state S5 and HDMI plugged in, while it can work when system resume.

Signed-off-by: Tommy Chung <tommy.chung@quanta.corp-partner.google.com>
Change-Id: I67d1af60b74343611e0ca6f54773216cf4caa895