TCPMv2: Refactor retimer update for USB MUX task
Now that mux sets may be completed in the USB_MUX task on boards which
enable it, refactor the retimer update mux get to wait when sets are
pending.  When sets are pending, the HOOKS task can come in later and
fill in the last_result with the mux state.

Additionally, set and validate cur_port any time an op comes in and is
then processed by the PD task.

BRANCH=None
BUG=b:208355638
TEST=on brya, ensure retimer FW update passes

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: Ied8ed1d90d1a2ea90b79ff4021ed77b858a65431