cr50: Revert U2F_Sign back to not checking authorization
In crrev.com/c/3221264 we uncommented the code that starts checking
auth_hmac for v1 key handles, but u2fd is not ready to provide the auth
secret yet. Comment the code back and put NULL for authTimeSecret for v1
key handles before secret enforement is implemented in u2fd.

BUG=b:210366574, b:172971998
TEST=make buildall -j

Change-Id: I8cf008213c88b8c88ab91f0601c319aea7ebfde0
Signed-off-by: Howard Yang <hcyang@google.com>