common/timer: Make clksrc_high volatile
This variable is updated from interrupt context (process_timers()
function) and read in get_time() which is called when CPU is not in
interrupt context.

Get_time() has mechanism to handle clksrc_high updates while getting
time. This actually works because there is a call to a function outside
the file between first clksrc_high read and second read in if statement.
Compiler doesn't know what function outside file does, it assumes that
clksrc_high could be updated (because process_timers() might be called)
so the variable is read twice actually, but innocent change in order of
lines could lead to unpredicted result.

To make sure that compiler does the right thing, add volatile keyword
in clksrc_high variable declaration.

BUG=b:200828093
BRANCH=none
TEST=make -j buildall

Signed-off-by: Patryk Duda <pdk@semihalf.com>
Change-Id: Ie701af387107bf76f07874d9eff1dc65fc844cf4