common/timer: Make clksrc_high volatile
This variable is updated from interrupt context (process_timers()
function, called from __hw_clock_source_irq()) and read in get_time()
which is called when CPU is not in interrupt context.

get_time() checks to see if clksrc_high was modified while obtaining
lower 32 bits of the timestamp (__hw_clock_source_read() call):

ts.le.hi = clksrc_high;
ts.le.lo = __hw_clock_source_read();
if (ts.le.hi != clksrc_high) {
  ts.le.hi = clksrc_high;
  ts.le.lo = __hw_clock_source_read();
}

Even though clksrc_high is not volatile, this works as-is because there
is a call to a function outside the file (__hw_clock_source_read)
between the first clksrc_high read and the second read in the if
statement. The compiler doesn't know what the function outside the file
does, so it assumes that clksrc_high could be updated. However, if the
code was changed so that it didn't have the external call between the
first clksrc_high read and the second, the compiler could optimize the
code and assume that clksrc_high was not changed, which would eliminat
the second read (and never execute the code in the if statement).

To make sure that compiler does the right thing, add volatile keyword
in clksrc_high variable declaration.

BUG=b:200828093
BRANCH=none
TEST=make -j buildall
TEST=Disassemble get_time() function using following command:
     arm-none-eabi-objdump --disassemble=get_time \
         ./build/bloonchipper/RW/ec.RW.elf

Result before applying patch:
080690cc <get_time>:
 80690cc:       b570            push    {r4, r5, r6, lr}
 80690ce:       4d07            ldr     r5, [pc, #28]
 80690d0:       4604            mov     r4, r0
 80690d2:       686e            ldr     r6, [r5, #4]
 80690d4:       f7f8 fcc3       bl      8061a5e <__hw_clock_source_read>
 80690d8:       686d            ldr     r5, [r5, #4]
 80690da:       42ae            cmp     r6, r5
 80690dc:       d001            beq.n   80690e2 <get_time+0x16>
 80690de:       f7f8 fcbe       bl      8061a5e <__hw_clock_source_read>
 80690e2:       e9c4 0500       strd    r0, r5, [r4]
 80690e6:       4620            mov     r0, r4
 80690e8:       bd70            pop     {r4, r5, r6, pc}
 80690ea:       bf00            nop
 80690ec:       200013c0        .word   0x200013c0

Result after applying patch:
080690cc <get_time>:
 80690cc:       b5f8            push    {r3, r4, r5, r6, r7, lr}
 80690ce:       4e07            ldr     r6, [pc, #28]
 80690d0:       4604            mov     r4, r0
 80690d2:       6877            ldr     r7, [r6, #4]
 80690d4:       f7f8 fcc3       bl      8061a5e <__hw_clock_source_read>
 80690d8:       6875            ldr     r5, [r6, #4]
 80690da:       42af            cmp     r7, r5
 80690dc:       d002            beq.n   80690e4 <get_time+0x18>
 80690de:       6875            ldr     r5, [r6, #4]
 80690e0:       f7f8 fcbd       bl      8061a5e <__hw_clock_source_read>
 80690e4:       e9c4 0500       strd    r0, r5, [r4]
 80690e8:       4620            mov     r0, r4
 80690ea:       bdf8            pop     {r3, r4, r5, r6, r7, pc}
 80690ec:       200013c0        .word   0x200013c0

Please note that there is additional clksrc_high load before second call
to __hw_clock_source_read() (inside the if statement).

Signed-off-by: Patryk Duda <pdk@semihalf.com>
Change-Id: Ie701af387107bf76f07874d9eff1dc65fc844cf4