stm32/hwtimer32: Clear status register after counter is set
Setting counter if overflow bit is not cleared it could result in
clksrc_high incrementing when it shouldn't happen.

To avoid this problem, just clear status register when new value is
assigned to counter. To avoid race between updating counter and clearing
status, operation is performed when counters are disabled.

BUG=b:200828093
BRANCH=none
TEST=make -j buildall

Signed-off-by: Patryk Duda <pdk@semihalf.com>
Change-Id: Ief65ac86fcf5c164f99855cc08c96c2ceddb29ea