stm32/hwtimer: Clear status register after counter is set
Whenever a timer interrupt occurs, the status register is checked for
overflow and the overflow status is passed to process_timers(), which
increments clksrc_high. When setting the clock to a new time, we don't
want the previous overflow state to trigger this process.

To avoid this problem, just clear status register when new value is
assigned to counter. To avoid race between updating counter and clearing
status, operation is performed when counters are disabled.

BUG=b:200828093
BRANCH=none
TEST=make -j buildall

Signed-off-by: Patryk Duda <pdk@semihalf.com>
Change-Id: Ifa3f2ff293483e167b5c7b6ff136708a876e8054