stm32/hwtimer32: Return UINT32_MAX event timestamp when it's not enabled
When timer is near overflow (next events are scheduled after overflow)
then matching next event timestamp is disabled. In IDLE task, EC is
taking that value and uses it to calculate time to next event. The
problem is that event timestamp can be in the past, so time to next
event becomes negative.

If all events are scheduled after overflow, then our next event is
overflow actually (when timestamp matching is set). The problem is
fixed by returning UINT32_MAX when timestamp matching is disabled.

BUG=b:200828093
BRANCH=none
TEST=make -j buildall
TEST=Run EC on bloonchipper and enable deep sleep. Check if device
     crashes.

Signed-off-by: Patryk Duda <pdk@semihalf.com>
Change-Id: Iaaf8eac8695214369954a10a922a378bfa01112a