zephyr: use node name for gpio and pwm structures
Use the node name instead of the label property for GPIO and PWM struct
names.

We currently define both a node name and label for named GPIO and PWM
nodes, which are normally set to match the net name, just lower and
upper case.

Switching the shim drivers to use the node name instead of the property
label in the internal data structures, so that we can eventually drop
the label and make the dts files a bit more compact.

This has the side effect of defining making the name lowercase, as the
current device tree API does not have a matching _UPPER name. For PWM it
should not matter since the name is only used in print logs, but for
GPIOs the name is reported by a host and console command, and while
find_signal_by_name() is case insensitive, this will change the name
reported by gpio_get_name().

If this turns out to cause problems (such as for some integration test
that depends on the net name), we could either fix those or revert this
CL, introduce DT_NODE_FULL_NAME_UPPER support in mainline Zephyr, and
then retry with that instead.

BRANCH=none
BUG=b:208436570
TEST=build and run on volteer

Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>
Change-Id: I594c2b571c4dc351476e49c106ed7b59eab37f24