gitlab: Download twister coverage from codecov.io
Loop over the merged commits from the upstream zephyr repo until we find
one that has a coverage report uploaded, and download that.

The coverage won't totally match up with the code we have checked out,
but it should be close, and this will save us many hours to re-run all
the tests.

BRANCH=None
BUG=None
TEST=Pushed manually to gitlab coverage branch

Signed-off-by: Jeremy Bettis <jbettis@google.com>
Change-Id: I2d2c4462e94136ae5e521b1b212e914e7c302550