zephyr/usbc: require MATH_UTIL to enable USB-C
The TCPM implementation depends on the bitmask_uint64 function in
multiple locations, so it should depend on the option that causes that
function to be built.

Since MATH_UTIL should have no effect on code size in most
configurations (especially with LTO enabled), its default is set to y in
order to simplify the configuration.

BUG=b:211693800
TEST=zmake testall; math_util is now built when PLATFORM_EC_USBC is
     enabled and MATH_UTIL is unconfigured
BRANCH=None

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: If3b18d43bd2e2cb21b22f57f298f942eadc2a963