pinmap: Remove interrupts, add pull up/down
Remove interrupt inputs. Previously, pins could be identified
as interrupts, which were considered inputs. Interrupts could be
identified as being triggered on rising, falling or both edges.
However, for the purposes of generating the DTS, interrupts are
considered exactly the same as inputs - the interrupt configuration is
treated completely separately. So there is no need to separately
deal with interrupts, they can all be made normal inputs.

Add pull up and pull down inputs. This is required so that the
hardware configuration can be correctly done for each GPIO.

BUG=none
TEST=go test ./...
BRANCH=none

Signed-off-by: Andrew McRae <amcrae@google.com>
Change-Id: I03082a1fe21060a5c0eaf421ffb8d806da8a83d0