npcx: gpio: add a new guard to KSI's GPIO_INT
In npcx chips, KSI pins are muxed with GPIO pins. In the current gpio
driver, the GPIO interrupt for these pins is enabled only when KEYSCAN
task is not defined. However, this assumption is not true when the ADC
anti-ghost feature is introduced. This CL adds a new guard for the
GPIO_INT of KSIs. The board level driver can define this flag to
override the default behavior.

BRANCH=none
BUG=b:213412432
TEST=pass "make buildall"

Signed-off-by: Jun Lin <CHLin56@nuvoton.com>
Change-Id: I21b2278f3bcf4b8bc685f5c2d7a4a4a1a0b7705b