zephyr: adc: switch to using io-channels
There is a special property named "io-channels" which can be used to
specify ADC module and channel.

Switch to use the "io-channels" property in the named-adc-channels node
instead of custom "channel" property.

The "io-channels" also specifies the ADC module, so there is no more
need of choosing "cros-ec,adc" in dts files.

BUG=b:213210342
TEST=zmake testall && make sure the 'adc' console command shows correct
values
BRANCH=main
Signed-off-by: Dawid Niedzwiecki <dn@semihalf.com>

Change-Id: I8b5b6a46ad6141fb0446cb9051cfd4eaa0c732fe