zephyr: docs: Cleanup notes about updating the chroot
This was written before Zmake's re-exec logic, which was the primary
reason we cautioned people to update the chroot or manually install
zephyr-build-tools so often.

Today, an update_chroot is still necessary for a toolchain update, but
this happens far less often and does not need noted here.

BUG=b:197636145
BRANCH=none
TEST=view in gitiles

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: I8e06ffa38f3501f5f2da5e390e3d55c3b1a5b3f8