zephyr: zmake: Add a README.md file generated from command structure
Generate a readme file from the command structure and arguments.  This
file gets checked that it matches the generated contents in the commit
queue using a new command "zmake generate-readme --diff".

Right now this file only has the commands, but could possibly be
extended in the future, for example, to document `BUILD.py` functions.

BUG=b:180609783
BRANCH=none
TEST=View readme in gitiles
TEST=Added unit tests

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: Ie8ed39b30ce2a58c91b2cf8e48b0426fb9b4f2e5