nissa/nivviks: configure USB-C1 chips only if present
To ensure port C1 is only enabled if actually present, leave it
unconfigured by default and configure the chips on the port only if a
type-C sub-board is actually present.

I've observed port C1 init messages when running on a system with no
configured sub-board; this should prevent that which may also be causing
memory corruption.

BUG=b:213996924
TEST=builds
BRANCH=none

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: I0cf89da06e6d632d7c2ed13818d2679c3282c6f5