zephyr: kingler: move GPIO_EN_USB_C1_SINK control to rt1718s hook
PPC NX20P3483 needs EN_SNK asserted to sink. This isn't a issue
on most cases except for the dead battery case. When dead battery
happens, the nx20p3483_init tries to enable sink from VCHG, but however,
the EN_SNK was controlled by the RT1718S's universal GPIO, and the
control was tight to charge_manager:board_set_active_charger_port().
The charge_manager wasn't ready at the very beginning moment of boot
up, and also the NX20P3483 won't response to I2C command when it was
in dead battery mode, so it has difficulty to interact to charger_manager
from PPC driver. So when NX20P3483 exits dead battery mode, due to the
EN_SNK was not asserted, the board is browned-out.

This CL moves the control from charge_manager to TCPC driver so
it's easier for NX20P3483 to ask RT1718S to enable EN_SNK pin.

BUG=b:214146359
TEST=Boot Kingler without battery with sinking from C1.
BRANCH=none

Change-Id: I032c020a958bd9d1cec1a6b10766c90ebd102f3b
Signed-off-by: Eric Yilun Lin <yllin@chromium.org>