Revert "make: disable kconfig option check"
Unfortunately this was disabled but never restarted. We have a number
of CONFIG options that are missing Kconfigs:

For volteer the list is:

   CONFIG_BATTERY_LOW_VOLTAGE_TIMEOUT
   CONFIG_HOSTCMD_DEBUG_MODE
   CONFIG_LED_PWM_COUNT
   CONFIG_POWER_S4_RESIDENCY
   CONFIG_STEINHART_HART_3V3_30K9_47K_4050B
   CONFIG_USB_PD_FLAGS

This cannot be applied until the above are fixed. Possibly this is not a
big problem anymore, since we are likely not adding much new
functionality to ECOS.

This reverts commit b3f5db5be78e15290dc7d80eaeecbb0f1b754a2c

BUG=b:195718112
BRANCH=none
TEST=make buildall -j32

Change-Id: I4b148caa04509721db83d001eb65dc061b2f21a6
Signed-off-by: Simon Glass <sjg@chromium.org>