anx3443: power down in S3 if the mux is in usb2 only mode
To save S3 power consumption when USB2 device (e.g. Gnubby) attached,
add a mechanism to detect if the device is USB2 or USB3.

If it's USB2 only, cache the mux state when entering S3, and restore
when back to S0.

Also removed the "Mux is not powered in Z1" related code since it's not
true.

BUG=b:209533566
TEST=see b:209533566 comment8
BRANCH=cherry

Signed-off-by: Ting Shen <phoenixshen@google.com>
Change-Id: Ib05dfe7868e57d3507c95f0020f94fd352d0034b