anx3443: power down in S3 if the mux is in usb2 only mode
To save S3 power consumption when USB2 device (e.g. Gnubby) attached,
add a mechanism to detect if the device is USB2 or USB3.

If it's USB2 only, cache the mux state when entering S3, and restore
when back to S0.

Also removed the "Mux is not powered in Z1" related code since it's not
true.

BUG=b:209533566
TEST=measure S0/S3 pp3300_z2 power in following scenario
     1) port 1 detached
       -> mux in power off mode
     2) port 1 connected to a C-to-A dongle
       -> mux in power off mode when s3
     3) port 1 connected to a type-c hub with DP
       -> mux always on
BRANCH=cherry

Signed-off-by: Ting Shen <phoenixshen@google.com>
Change-Id: Ib05dfe7868e57d3507c95f0020f94fd352d0034b