zephyr: Avoid calling task_get_current except in tasks.
task_get_current is being called in in_deferred_context(), which
may be invoked in initialisation code before the tasks
have been started (so shimmed_tasks_dyn was not initialised).

This fix avoids the use of any data that requires the EC tasks
to have been initialised or started.

BUG=b:213996924
TEST=zmake configure -b nivviks; flash and run
BRANCH=none

Signed-off-by: Andrew McRae <amcrae@google.com>
Change-Id: I0bd1a13a8d5bbf8b3204695b7f9977fe99f63437