zephyr: zmake: Force clobbering on DTS or kconfig file changes
Zephyr's build system has incorrecly specified cmake dependencies, and
may not re-run cmake when necessary (for example, due to kconfig or
device tree changes).

Ultimately, it would be preferred if this can be fixed
upstream (b/215560602 tracks that effort internally).  However, it is
easy enough to work around this in zmake and force clobbering when we
detect the condition.  Do this as a temporary measure so that people
don't start habitually passing --clobber, even when it's unnecessary.

BUG=b:215560602
BRANCH=none
TEST=updated unit tests

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: Id581ccedf290cd21b31241673318ef5745678767