zephyr test: Configure charging power in DTS
Use the device tree to configure the battery's desired charging voltage
and current. Do not modify the device parameters during the driver unit
test. The battery charging voltage and current are not directly related
to the VBUS voltage and current, so do not check that the charging
voltage matches VBUS.

BUG=b:213908743,b:213909940
TEST=zmake configure --test zephyr/test/drivers
BRANCH=none

Signed-off-by: Abe Levkoy <alevkoy@chromium.org>
Change-Id: I99ca3da7e49755587cd73347bb56f5e372e99d91