zephyr: Fix named-gpios et al to drop #gpio-cells
The #gpio-cells property is only required when the referenced node
is a GPIO controller. named-gpios is not a GPIO controller, and
cannot be used as a GPIO reference.
DTS and code can refer to named-gpios directly, and then retrieve
the GPIO from the named-gpio gpios property.

Also rename irq-gpio to irq-pin so that DTS does not complain
about missing #gpio-cells.

BUG=b:216402316
TEST=zmake testall
BRANCH=none

Signed-off-by: Andrew McRae <amcrae@google.com>
Cq-Depend: chromium:3418800, chromium:3418799, chromium:3423294
Change-Id: Ifdea8662647c092950962d3b6b0f32acda23aa5a