zephyr: remove HOST_EVENT_MASK macro
This macro was introduced as workaround to differences in host events
definitions between CrOS EC and Zephyr. Since zephyr macros define
masks, instead of offsets, it will be easier to define _NONE as 0,
instead of BIT(0) and rest of the fields as bit masks.
It will remove the differences between host-event-wakeup-mask and
event-wakeup-mask definitions in device tree. They use the same
compatible string, so there shouldn't be requirement about using
macro in only one of the definitions.

BUG=none
BRANCH=main
TEST=Same test as in commit that introduced this macro:
     Suspend DUT with powerd_dbus_suspend, and make sure the EC command
     "hostevent set 0x0000000002000000" wakes up the board.

Change-Id: Ie582ecd805b1999d09ec71b10cdeff0b5ac1ec82
Signed-off-by: Michał Barnaś <mb@semihalf.com>