qcom: Move the execution of warm reset to CHIPSET context
In CrOS EC, the `apreset` command is executed in the CONSOLE
context. But in Zephyr, the command is executed in the lowest
priority task context. This low priority task is easily
preempted and makes the power signal missing. This miss is
wrongly treated as the PMIC not respond correctly and trigger
a cold reset sequence: S0 -> S5 -> S0.

This CL moves the execution to the CHIPSET task context. It
is now independent from the console runtime.

BRANCH=None
BUG=b:215252361
TEST=Call the `apreset` command, just a warm reset triggered.

Change-Id: I2be65db46b453421c2a04d10a75c01dbf16f1f89
Signed-off-by: Wai-Hong Tam <waihong@google.com>