zephyr: npcx: refactor the calculation of total RAM blcok
The calculation of the total RAM blocks is determined by the size of
the code RAM, data RAM, and the RAM used by the booter. In the current
source, the RAM used by the booter is hardcoded. This CL replaces the
hardcoding by fetching the size from the DT node bt_ram.

BUG=none
BRANCH=none
TEST="zmake testall -clobber"
TEST="check the total RAM block is correct for different npcx variants"

Signed-off-by: Jun Lin <CHLin56@nuvoton.com>
Change-Id: Ia04710360ec4d3144860fb1b33763e2230747c44