zephyr: pinmap: handle a new INPUT_L pin type
Some signals are logically high when the pin is low e.g. WP is active,
when the ec_wp_odl pin is low. There is a new pin type "INPUT_L" to
handle such GPIOs. Add GPIO_ACTIVE_LOW flag to pins with that type in
the generated DTS. It allows using Zephyr GPIO API to get the logical
state.

BUG=b:211779766
TEST=Generate a dts file and make sure the GPIO_ACTIVE_LOW flag is added
to pins with INPUT_L type.
BRANCH=main

Signed-off-by: Dawid Niedzwiecki <dn@semihalf.com>
Change-Id: I1c4ed074856019cf70394c72683b3ec00a1c5f03