zephyr: pinmap: add GPIO_ACTIVE_LOW flag
Some signals are logicaly hight when the pin is low e.g. WP is active,
when the ec_wp_odl pin is low. Add GPIO_ACTIVE_LOW flag to GPIO in a
generated DTS that has "_L" at the end of enum name. It allows using
Zephyr GPIO API to get the logical state.

BUG=b:211779766
TEST=Generate a dts file and make sure the GPIO_ACTIVE_LOW flag is added
BRANCH=main

Signed-off-by: Dawid Niedzwiecki <dn@semihalf.com>
Change-Id: I1c4ed074856019cf70394c72683b3ec00a1c5f03