nissa: merge common kconfig options into prj.conf
It's easy to update configuration for one board variant without touching
the other when a given change is actually applicable to both. This
change merges the common kconfig options used by Nissa to avoid
duplicating them, which should help ensure the feature set for the two
variants doesn't diverge significantly. Where config is
expected to vary by board, they are not merged.

The output changes are:
 * CONFIG_FPU is left at default (=y) for Nivviks, rather than enabled
   explicitly
 * PSL is disabled on nivviks because it is unused (EC wakeup is handled by
   the external Z-state circuitry on the mainboard)
 * PWM shell is enabled on nivviks
 * Sensor support is enabled on nereid

The board_is_sourcing_vbus function is also removed from nereid because
it was unused: it is only called when
CONFIG_PLATFORM_EC_USB_PD_5V_EN_CUSTOM is enabled, which it is not on
nereid.

BUG=none
TEST=diff of old .config against new only shows expected changes
BRANCH=none

Change-Id: I02d61a8d71ff730c442e60fa2e1ef24ff46c7095
Signed-off-by: Peter Marheine <pmarheine@chromium.org>