nissa: merge some kconfig options into prj.conf
It's easy to update configuration for one board variant without touching
the other when a given change is actually applicable to both. This
change merges some of the kconfig options used by Nissa to avoid
duplicating them, which should help ensure the feature set for the two
variants doesn't diverge significantly.

This does not yet merge all relevant configuration items, but does so
for a set of core features that should be uncontroversial: they
represent features that I don't expect would ever be disabled for a
system which could still be called a Nissa variant.

BUG=none
TEST=TODO: diff .config
BRANCH=none

Change-Id: I02d61a8d71ff730c442e60fa2e1ef24ff46c7095
Signed-off-by: Peter Marheine <pmarheine@chromium.org>