util: Allow ignoring problematic Kconfig options
The script normally complains if an option is in config_allowed but is
actually defined in Kconfig.

This doesn't always work though, since it is possible that the option
is defined in Zephyr itself, not in the platform/etc/zephyr directory.

Add an option to ignore them.

BUG=b:181323955
BRANCH=none
TEST=./util/kconfig_check.py -c build/blipper/.config
   -a util/config_allowed.txt -p PLATFORM_EC_ -s zephyr/
   -I ~/cosarm/src/third_party/zephyr/main -i DAC check
See that this does not cause an error now

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: Icbb245b0cd4dcdc59a72a970020314d8a265fb98