zephyr: Do not call power_set_state in tests
Change tests to not call power_set_state() as it doesn't actually change
the power state, since the chipset_task only reads the state once and
then keeps the state in a local var.

There are a few exceptions to this. Some tests run before chipset_task
is started, those continue to use power_set_state(). Also the hibernate
tests make some redundant calls to set state to G3 when it already is
in that state, just to reset the last_shutdown_time. Otherwise the
g3 time will be unpredictable.

Fixed test_set_chipset_to_s0 to actually coax the chipset_task to go to
S0.

BRANCH=None
BUG=b:214256453
TEST=zmake testall

Signed-off-by: Jeremy Bettis <jbettis@google.com>
Change-Id: I04ea6f73fcf784f20e53049eb3276173efd2655f