intel_x86: Add config to configure EC S5 exit timer
On enabling Delayed Authentication Mode(DAM) in the coreboot for latest
platforms, system takes more time to move from G3 to S5 due to which EC
G3 to S5 timer gets timedout and EC forces the system to G3. This
patches adds a new config to configure the timeout value.

BRANCH=None
BUG=b:191742284
TEST=Boot test on Brya with the DAM enabled coreboot

Signed-off-by: Sugnan Prabhu S <sugnan.prabhu.s@intel.com>
Change-Id: I78821280fa5928682c9c0c742a4e00299ac49db0
Signed-off-by: Sugnan Prabhu S <sugnan.prabhu.s@intel.com>