intel_x86: Add config to configure S5 exit timer
On enabling Delayed Authentication Mode(DAM) in the coreboot for latest
platforms, system takes more time to exit from S5 than the current wait
time(4secs). Due to which EC forces the system to G3 after few retries.
This patch adds a new config to allow configuration of this timeout
value for which EC has to wait for S5 exit before performing RTC reset
and forcing the system to G3.

BRANCH=None
BUG=b:191742284
TEST=Boot test on Brya with the DAM enabled coreboot

Signed-off-by: Sugnan Prabhu S <sugnan.prabhu.s@intel.com>
Change-Id: I78821280fa5928682c9c0c742a4e00299ac49db0