cq: Handle coverage better
Only build/test with coverage if --code-coverage is passed.

Change bundle_coverage to merge all lcov files together, and bundle a
single lcov.info that includes all boards and all zephyr tests. This is
different from what we do on gitlab, but I don't want to build the
CROS-EC tests here.  For purposes of showing coverage in gerrit, this
should be good enough.

BRANCH=None
BUG=b:156895937
TEST=Ran commands in README.md, and it caught the zmake breakage at tot

Change-Id: I517b191d097384da59fee4c27cffcfc58b7d20ec
Signed-off-by: Jeremy Bettis <jbettis@google.com>