stm32: Add the DFU Runtime identifiers
Adds a DFU Runtime identifier and a DFU bootmanager to control the
switch between the DFU and main application region.

The DFU runtime identifier operates on the main application in the RW
region. A DFU host recognizes the interface and can send commands to
switch the firmware into DFU mode.

BRANCH=servo
BUG=b:217955677
TEST=Enabled the DFU configuration on Servo_v4 with CONFIG_DFU_RUNTIME
TEST='sudo dfu-util -l' Verified the dfu-util recognizes the
    DFU runtime device.

Signed-off-by: Brian Nemec <bnemec@chromium.org>
Change-Id: I0e02b7cb35a7c594f169c56a96f965e5d87e3cfb