stm32: Add the DFU Runtime identifiers and DFU Bootmanager
Adds a DFU Runtime identifier and a DFU bootmanager to control the switch
between the DFU and main application region.

The DFU runtime identifier operates on the main application in the RW
region. A DFU-host recognizes the interface and can send commands to
send the firmware into DFU mode. Instead of jumping to DFU directly,
the software sets backup flag registers and jumps into the
DFU Bootmanager in the RO region which performs the actual jump to DFU.

The DFU Bootmanager performs self checks when the application boots, the
result of these checks will determine if the system jumps to RW or DFU, if
any of the following conditions are true it will jump into DFU:

* The system backup registers indicate the 'jump to DFU' request has been sent.
* The start of the RW region is erased indicating the prior firmware update
    has failed.
* An optional check can be enabled to record the number of consecutive
    unexpected reboots which will be useful for developer test images to avoid
    locked systems.

Testing requires enabling multiple configuration settings set in the
follower CL which enables it on Servo_v4. dfu-util version 0.9 lacks
the 'leave' parameter required to automatically jump out, systems using
this would need to be power cycled to leave DFU.

BRANCH=servo
BUG=b:217955677
TEST=Enabled the DFU configuration on a Servo_v4. Verified that the
    runtime identifier is correctly recognized by dfu-util and pydfu.py
    implementations of the DFU host protocol. Verified firmware updates
    via the DFU protocol function correctly.
TEST=sudo dfu-util -a 0 -s 0x08000000:leave -e -D ec.bin

Signed-off-by: Brian Nemec <bnemec@chromium.org>
Change-Id: I0e02b7cb35a7c594f169c56a96f965e5d87e3cfb