lazor: Use npcx7m7fc instead of npcx7m6fc
In spite of the fact that lazor has a npcx7m6fc EC,
force it to use npcx7m7fc so that the extra code RAM
is accessible. These parts are the same die, just binned
separately, and the extra RAM is not tested.
However enabling the options required for zephyr requires
the extra space.

BUG=b:218401333
TEST=zmake configure -b lazor
BRANCH=none

Signed-off-by: Andrew McRae <amcrae@google.com>
Change-Id: Ia24415d8521da4ed894f3feaf6568410526e2f91