zephyr: npcx: add build time check for the hibernate DT node
Nuvoton EC supports two hibernate mode:
 - PSL mode
 - non-PSL (disable RAM) mode

The PSL mode uses the DT node vsby-psl-in-list to configure the PSL
input wakeup pins. The non-PSL mode uses the DT node hibernate-wake-pins
to configure the GPIO wakeup pins. This CL adds a build time assert to
prevent a user from using the incorrect DT node for different mode.

Also, add status = "okay" to the vsby-psl-in-list node because the
status was disabled at default with PR:42668.

BUG=b:216893756
BRANCH=none
TEST=set "zmake testall --clobber"
TEST=add "CONFIG_PLATFORM_EC_HIBERNATE_PSL=y" and define
hibernate-wake-pins node in the prj.conf of npcx9_evb; see the build
error.
TEST=set "CONFIG_PLATFORM_EC_HIBERNATE_PSL=n" and define
vsby-psl-in-list node in the prj.conf of npcx9_evb; see the build
error.

Signed-off-by: Jun Lin <CHLin56@nuvoton.com>
Change-Id: I83f4c3d427ef479445d5e5b992f53056da9bf1ef